+++ server/src/main/java/org/opensearch/index/search/stats/SearchStats.java
@@ -142,6 +142,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
         private long queryCount;
         private long queryTimeInMillis;
         private long queryCurrent;
+        private long queryFailedCount;
 
         private long concurrentQueryCount;
         private long concurrentQueryTimeInMillis;
@@ -169,6 +170,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
         private long starTreeQueryCount;
         private long starTreeQueryTimeInMillis;
         private long starTreeQueryCurrent;
+        private long starTreeQueryFailed;
 
         @Nullable
         private RequestStatsLongHolder requestStatsLongHolder;
@@ -191,6 +193,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
             this.queryCount = builder.queryCount;
             this.queryTimeInMillis = builder.queryTimeInMillis;
             this.queryCurrent = builder.queryCurrent;
+            this.queryFailedCount = builder.queryFailedCount;
 
             this.concurrentQueryCount = builder.concurrentQueryCount;
             this.concurrentQueryTimeInMillis = builder.concurrentQueryTimeInMillis;
@@ -218,6 +221,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
             this.starTreeQueryCount = builder.starTreeQueryCount;
             this.starTreeQueryTimeInMillis = builder.starTreeQueryTimeInMillis;
             this.starTreeQueryCurrent = builder.starTreeQueryCurrent;
+            this.starTreeQueryFailed = builder.starTreeQueryFailed;
         }
 
         /**
@@ -319,12 +323,18 @@ public class SearchStats implements Writeable, ToXContentFragment {
                 starTreeQueryTimeInMillis = in.readVLong();
                 starTreeQueryCurrent = in.readVLong();
             }
+
+            if (in.getVersion().onOrAfter(Version.V_3_3_0)) {
+                queryFailedCount = in.readVLong();
+                starTreeQueryFailed = in.readVLong();
+            }
         }
 
         public void add(Stats stats) {
             queryCount += stats.queryCount;
             queryTimeInMillis += stats.queryTimeInMillis;
             queryCurrent += stats.queryCurrent;
+            queryFailedCount += stats.queryFailedCount;
 
             concurrentQueryCount += stats.concurrentQueryCount;
             concurrentQueryTimeInMillis += stats.concurrentQueryTimeInMillis;
@@ -352,11 +362,13 @@ public class SearchStats implements Writeable, ToXContentFragment {
             starTreeQueryCount += stats.starTreeQueryCount;
             starTreeQueryTimeInMillis += stats.starTreeQueryTimeInMillis;
             starTreeQueryCurrent += stats.starTreeQueryCurrent;
+            starTreeQueryFailed += stats.starTreeQueryFailed;
         }
 
         public void addForClosingShard(Stats stats) {
             queryCount += stats.queryCount;
             queryTimeInMillis += stats.queryTimeInMillis;
+            queryFailedCount += stats.queryFailedCount;
 
             concurrentQueryCount += stats.concurrentQueryCount;
             concurrentQueryTimeInMillis += stats.concurrentQueryTimeInMillis;
@@ -381,6 +393,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
 
             starTreeQueryCount += stats.starTreeQueryCount;
             starTreeQueryTimeInMillis += stats.starTreeQueryTimeInMillis;
+            starTreeQueryFailed += stats.starTreeQueryFailed;
         }
 
         public long getQueryCount() {
@@ -397,6 +410,10 @@ public class SearchStats implements Writeable, ToXContentFragment {
 
         public long getQueryCurrent() {
             return queryCurrent;
+        }
+
+        public long getQueryFailedCount() {
+            return queryFailedCount;
         }
 
         public long getConcurrentQueryCount() {
@@ -507,6 +524,10 @@ public class SearchStats implements Writeable, ToXContentFragment {
             return starTreeQueryCurrent;
         }
 
+        public long getStarTreeQueryFailed() {
+            return starTreeQueryFailed;
+        }
+
         public static Stats readStats(StreamInput in) throws IOException {
             return new Stats(in);
         }
@@ -562,6 +583,11 @@ public class SearchStats implements Writeable, ToXContentFragment {
                 out.writeVLong(starTreeQueryTimeInMillis);
                 out.writeVLong(starTreeQueryCurrent);
             }
+
+            if (out.getVersion().onOrAfter(Version.V_3_3_0)) {
+                out.writeVLong(queryFailedCount);
+                out.writeVLong(starTreeQueryFailed);
+            }
         }
 
         @Override
@@ -569,6 +595,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
             builder.field(Fields.QUERY_TOTAL, queryCount);
             builder.humanReadableField(Fields.QUERY_TIME_IN_MILLIS, Fields.QUERY_TIME, getQueryTime());
             builder.field(Fields.QUERY_CURRENT, queryCurrent);
+            builder.field(Fields.QUERY_FAILED_TOTAL, queryFailedCount);
 
             builder.field(Fields.CONCURRENT_QUERY_TOTAL, concurrentQueryCount);
             builder.humanReadableField(Fields.CONCURRENT_QUERY_TIME_IN_MILLIS, Fields.CONCURRENT_QUERY_TIME, getConcurrentQueryTime());
@@ -578,6 +605,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
             builder.field(Fields.STARTREE_QUERY_TOTAL, starTreeQueryCount);
             builder.humanReadableField(Fields.STARTREE_QUERY_TIME_IN_MILLIS, Fields.STARTREE_QUERY_TIME, getStarTreeQueryTime());
             builder.field(Fields.STARTREE_QUERY_CURRENT, getStarTreeQueryCurrent());
+            builder.field(Fields.STARTREE_QUERY_FAILED, getStarTreeQueryFailed());
 
             builder.field(Fields.FETCH_TOTAL, fetchCount);
             builder.humanReadableField(Fields.FETCH_TIME_IN_MILLIS, Fields.FETCH_TIME, getFetchTime());
@@ -633,6 +661,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
             private long queryCount = 0;
             private long queryTimeInMillis = 0;
             private long queryCurrent = 0;
+            private long queryFailedCount = 0;
             private long concurrentQueryCount = 0;
             private long concurrentQueryTimeInMillis = 0;
             private long concurrentQueryCurrent = 0;
@@ -653,6 +682,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
             private long starTreeQueryCount = 0;
             private long starTreeQueryTimeInMillis = 0;
             private long starTreeQueryCurrent = 0;
+            private long starTreeQueryFailed = 0;
             @Nullable
             private RequestStatsLongHolder requestStatsLongHolder = null;
 
@@ -670,6 +700,11 @@ public class SearchStats implements Writeable, ToXContentFragment {
 
             public Builder queryCurrent(long current) {
                 this.queryCurrent = current;
+                return this;
+            }
+
+            public Builder queryFailed(long count) {
+                this.queryFailedCount = count;
                 return this;
             }
 
@@ -770,6 +805,11 @@ public class SearchStats implements Writeable, ToXContentFragment {
 
             public Builder starTreeQueryCurrent(long current) {
                 this.starTreeQueryCurrent = current;
+                return this;
+            }
+
+            public Builder starTreeQueryFailed(long count) {
+                this.starTreeQueryFailed = count;
                 return this;
             }
 
@@ -916,6 +956,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
         static final String QUERY_TIME = "query_time";
         static final String QUERY_TIME_IN_MILLIS = "query_time_in_millis";
         static final String QUERY_CURRENT = "query_current";
+        static final String QUERY_FAILED_TOTAL = "query_failed";
         static final String CONCURRENT_QUERY_TOTAL = "concurrent_query_total";
         static final String CONCURRENT_QUERY_TIME = "concurrent_query_time";
         static final String CONCURRENT_QUERY_TIME_IN_MILLIS = "concurrent_query_time_in_millis";
@@ -925,6 +966,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
         static final String STARTREE_QUERY_TIME = "startree_query_time";
         static final String STARTREE_QUERY_TIME_IN_MILLIS = "startree_query_time_in_millis";
         static final String STARTREE_QUERY_CURRENT = "startree_query_current";
+        static final String STARTREE_QUERY_FAILED = "startree_query_failed";
         static final String FETCH_TOTAL = "fetch_total";
         static final String FETCH_TIME = "fetch_time";
         static final String FETCH_TIME_IN_MILLIS = "fetch_time_in_millis";